Initial version of the site
This commit adds the initial version of the site, including the database and
table generation, submissions page, and about page. Most of the content is a
replica from the old site
(https://web.archive.org/web/20220327112752/https://www.canceledpeople.com/) to
keep it consistent. A subsequent commit will modify the about page to mention
that this is not the old site, but is a revived version of it under new
management since the old one mysteriously went missing, but is a global
treasure.

As of now, the site does not use any build process or page generation step, and
is written in raw HTML, CSS, JS. Probably some lightweight page build step will
be added in the future. Also the site doesn't work in browsers that don't
support JS import assertions, which is a problem.